Sync input text with redux
This is implemented as a fully controlled text input dispatching on
every keystroke. That's not how I usually do things because it tends to
damage performance and clutter the action logs, but my main goal ATM is
shipping a demo.

Usually I'll create an intermediate `<Input>` element that switches
between controlled and uncontrolled modes in response to focus states
and only dispatch on meaningful change events. I may add the component
here, or I may publish it separately. Perhaps it already exists as
a library.

Here's an example of what I'm referencing:
https://github.com/PsychoLlama/ponder/blob/2e22f7601c50de5650e8d33fd1acd9b2e5570048/packages/ponder-ui/src/Input.tsx