Rework redux model
I'd originally planned on keeping state as simple as possible and
closely mirroring the redux TodoMVC demo, but plans are meant to be
broken! Well, to be more specific, I ran into the same problems I always
run into when representing entities in arrays.

The list view needs a collection of every todo item ordered ascending by
timestamp, along with computed counts. For each todo, render a sub-item
so React is responsible for keeping track of the ID (rather than hacking
it into component state).

Passing an index as an ID is frustratingly brittle. Asynchronous work is
very dangerous, particularly animations, since it obviously isn't
a static reference to the entity in question and you may end up
inadvertently affecting the wrong entity. The alternative is passing
state directly down to the child, but if the child requires any other
redux state, the duplication of selection reveals just how obviously
dependent your global state is on the component hierarchy (otherwise
you'd combine the selectors in the same file). No, passing anything
other than IDs or flags down the component hierarchy is symptomatic of
a poorly constructed state.

Replacing the task list with a task map allows me to pass around IDs
more freely. This is a pattern I use quite often in proprietary
projects.