Set homepage path
I've hosted the project on GH Pages which by default appends
a `/<project>` path to the hosted assets, but CRA builds under the
assumption of a root domain. The JS assets are failing to load because
they're requesting the wrong path. Setting the application route helps
CRA know which path to hit.