Isolate new task title in component state
TL;DR: `state.newTaskTitle` moved into component state,
`tasks.create(...)` takes a title now.

This is a continuation of b82f236. I published the optimized form
element as `freeform-input` on npm, hacked it into this app, then
quickly realized it wouldn't work. So long as the user stayed focused on
the input, there was no way to clear the input for the next todo. That
was by design, although an unhappy coincidence. On the bright side,
I realized it's perfect for a related feature: double click to edit!

So rather than depend on a fancy new component, I've moved controlled
input state out of redux. No more controlled input dispatches cluttering
the logs. Even better, I'm able to set the task title through the
`tasks.create(...)` instead of communicating it through state. Of
course, that only has benefit when you're using actions outside the
component tree, a true rarity in React apps.

The real value here is less dispatches.