Only deploy from master
Now that the project is a bit more mature, I'm going through and
cleaning up the obvious rough edges. One thing that was "concerning" was
that GH Pages deployed from pull requests. That's, uh, that's not good.

This splits out the deploy process into another job that exclusively
targets pushes to master.

I also snuck in a GH action that should help with install times by
caching the node modules folder. That matters more, now that electron is
part of the testing process.